hba1c changed to double

+++ app/src/main/java/com/heartcare/agni/data/local/roomdb/entities/vitals/VitalEntity.kt
@@ -39,7 +39,7 @@ data class VitalEntity(
 
     @Embedded(prefix = "hipCircumference_")
     val hipCircumference: Measurement?,
-    val hbA1cPercentage: Int?,
+    val hbA1cPercentage: Double?,
 
     @Embedded(prefix = "serumCreatinine_")
     val serumCreatinine: Measurement?,
